+++ api/package.json
@@ -30,16 +30,19 @@
     "@nestjs/config": "^3.1.1",
     "@nestjs/core": "^10.0.0",
     "@nestjs/cqrs": "^10.2.6",
+    "@nestjs/passport": "^10.0.3",
     "@nestjs/platform-express": "^10.3.3",
     "@nestjs/swagger": "^7.2.0",
     "@nestjs/terminus": "^10.2.0",
     "@nestjs/typeorm": "^10.0.1",
-    "@supabase/supabase-js": "^2.40.0",
+    "@supabase/supabase-js": "^2.43.1",
     "@types/multer": "^1.4.11",
     "class-transformer": "^0.5.1",
     "class-validator": "^0.14.0",
     "fred-api": "^2.0.1",
     "ioredis": "^5.3.2",
+    "passport": "^0.7.0",
+    "passport-jwt": "^4.0.1",
     "pg": "^8.11.3",
     "reflect-metadata": "^0.1.13",
     "rxjs": "^7.8.1",
