+++ api/src/auth/auth.module.ts
@@ -1,8 +1,8 @@
 import { Module } from '@nestjs/common';
-import { AuthService } from './auth.service';
-import { AuthController } from './auth.controller';
+import { AuthService } from './application/auth.service';
+import { AuthController } from './api/auth.controller';
 import { TypeOrmModule } from '@nestjs/typeorm';
-import { MemberEntity } from './member.entity';
+import { MemberEntity } from './entity/member.entity';
 
 @Module({
   imports: [TypeOrmModule.forFeature([MemberEntity])],
