+++ api/src/auth/entity/member.entity.ts
