+++ api/src/auth/util/get-member.decorator.ts
@@ -1,5 +1,5 @@
 import { createParamDecorator, ExecutionContext } from '@nestjs/common';
-import { MemberEntity } from './member.entity';
+import { MemberEntity } from '../entity/member.entity';
 
 export const Member = createParamDecorator((data, ctx: ExecutionContext): MemberEntity => {
   const req = ctx.switchToHttp().getRequest();
