+++ api/src/numerical-guidance/infrastructure/adapter/persistence/custom-forecast-indicator/mapper/custom-forecast-indicator.mapper.ts
@@ -1,6 +1,6 @@
 import { CustomForecastIndicator } from 'src/numerical-guidance/domain/custom-forecast-indicator';
 import { CustomForecastIndicatorEntity } from '../entity/custom-forecast-indicator.entity';
-import { MemberEntity } from 'src/auth/member.entity';
+import { MemberEntity } from 'src/auth/entity/member.entity';
 
 export class CustomForecastIndicatorMapper {
   static mapDomainToNewEntity(
