+++ api/src/numerical-guidance/numerical-guidance.module.ts
@@ -7,8 +7,8 @@ import { CreateIndicatorBoardMetadataCommandHandler } from './application/comman
 import { IndicatorBoardMetadataPersistentAdapter } from './infrastructure/adapter/persistence/indicator-board-metadata/indicator-board-metadata.persistent.adapter';
 import { TypeOrmModule } from '@nestjs/typeorm';
 import { IndicatorBoardMetadataEntity } from './infrastructure/adapter/persistence/indicator-board-metadata/entity/indicator-board-metadata.entity';
-import { AuthService } from '../auth/auth.service';
-import { MemberEntity } from '../auth/member.entity';
+import { AuthService } from '../auth/application/auth.service';
+import { MemberEntity } from '../auth/entity/member.entity';
 import { GetIndicatorBoardMetadataQueryHandler } from './application/query/indicator-board-metadata/get-indicator-board-metadata/get-indicator-board-metadata.query.handler';
 import { InsertIndicatorIdCommandHandler } from './application/command/indicator/insert-indicator-id/insert-indicator-id.command.handler';
 import { GetIndicatorBoardMetadataListQueryHandler } from './application/query/indicator-board-metadata/get-indicator-board-metadata-list/get-indicator-board-metadata-list.query.handler';
