+++ api/src/numerical-guidance/test/E2E test/history-indicator/history-indicator.e2e.spec.ts
@@ -1,7 +1,7 @@
 import { HttpStatus, INestApplication, ValidationPipe } from '@nestjs/common';
 import { DataSource } from 'typeorm';
 import { IndicatorEntity } from '../../../infrastructure/adapter/persistence/indicator/entity/indicator.entity';
-import { MemberEntity } from '../../../../auth/member.entity';
+import { MemberEntity } from '../../../../auth/entity/member.entity';
 import { HistoryIndicatorEntity } from '../../../infrastructure/adapter/persistence/history-indicator/entity/history-indicator.entity';
 import * as fs from 'fs';
 import { HistoryIndicatorValueEntity } from '../../../infrastructure/adapter/persistence/history-indicator-value/entity/history-indicator-value.entity';
@@ -14,9 +14,9 @@ import { TypeOrmModule } from '@nestjs/typeorm';
 import { HttpModule } from '@nestjs/axios';
 import { HistoryIndicatorController } from '../../../api/history-indicator/history-indicator.controller';
 import { AdjustIndicatorValue } from '../../../util/adjust-indicator-value';
-import { AuthService } from '../../../../auth/auth.service';
+import { AuthService } from '../../../../auth/application/auth.service';
 import { GetHistoryIndicatorQueryHandler } from '../../../application/query/history-indicator/get-history-indicator/get-history-indicator.query.handler';
-import { AuthGuard } from '../../../../auth/auth.guard';
+import { AuthGuard } from '../../../../auth/util/auth.guard';
 import { of } from 'rxjs';
 import { HttpExceptionFilter } from '../../../../utils/exception-filter/http-exception-filter';
 import * as request from 'supertest';
