+++ api/src/numerical-guidance/test/integration-test/adapter/persistence/custom-forecast-indicator.persistent.adapter.spec.ts
@@ -4,8 +4,8 @@ import { ConfigModule, ConfigService } from '@nestjs/config';
 import { Test } from '@nestjs/testing';
 import { TypeOrmModule } from '@nestjs/typeorm';
 import { PostgreSqlContainer } from '@testcontainers/postgresql';
-import { AuthService } from 'src/auth/auth.service';
-import { MemberEntity } from 'src/auth/member.entity';
+import { AuthService } from 'src/auth/application/auth.service';
+import { MemberEntity } from 'src/auth/entity/member.entity';
 import { CustomForecastIndicator } from 'src/numerical-guidance/domain/custom-forecast-indicator';
 import { CustomForecastIndicatorPersistentAdapter } from 'src/numerical-guidance/infrastructure/adapter/persistence/custom-forecast-indicator/custom-forecast-indicator.persistent.adapter';
 import { CustomForecastIndicatorEntity } from 'src/numerical-guidance/infrastructure/adapter/persistence/custom-forecast-indicator/entity/custom-forecast-indicator.entity';
